requirements.txt
appdirs==1.4.4
audioread==2.1.9
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cycler==0.10.0
decorator==5.0.9
ffmpeg-python==0.2.0
future==0.18.2
idna==2.10
ImageHash==4.2.0
joblib==1.0.1
kiwisolver==1.3.1
librosa==0.8.1
llvmlite==0.36.0
numba==0.53.1
numpy==1.20.3
packaging==20.9
Pillow==8.2.0
pooch==1.3.0
pycparser==2.20
pydub==0.25.1
pyparsing==2.4.7
python-dateutil==2.8.1
PyWavelets==1.1.1
requests==2.25.1
resampy==0.2.2
scikit-learn==0.24.2
scipy==1.6.3
six==1.16.0
SoundFile==0.10.3.post1
threadpoolctl==2.1.0
urllib3==1.26.5
